Fix/fridge service refactoring

+++ src/main/java/com/recipe/app/src/fridge/FridgeController.java
@@ -15,7 +15,6 @@ import org.springframework.web.context.request.ServletRequestAttributes;
 
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
 import static com.recipe.app.config.BaseResponseStatus.*;
@@ -52,7 +51,6 @@ public class FridgeController {
     @PostMapping("/fridges")
     public BaseResponse<List<PostFridgesRes>> postFridges(@RequestBody PostFridgesReq parameters) {
         try {
-
             List<FridgeBasketList> fridgeBasketList = parameters.getFridgeBasketList();
             Integer userIdx = jwtService.getUserId();
 
@@ -60,54 +58,42 @@ public class FridgeController {
                 return new BaseResponse<>(POST_FRIDGES_EMPTY_FRIDGE_BASKET_LIST);
             }
 
+            for (FridgeBasketList fridgeBasket : fridgeBasketList) {
+                String ingredientName = fridgeBasket.getIngredientName();
+                String ingredientIcon = fridgeBasket.getIngredientIcon();
+                Integer ingredientCategoryIdx = fridgeBasket.getIngredientCategoryIdx();
+                String expiredAt = fridgeBasket.getExpiredAt();
+                String storageMethod = fridgeBasket.getStorageMethod();
+                Integer count = fridgeBasket.getCount();
 
-            for (int i = 0; i < fridgeBasketList.size(); i++) {
-                String ingredientName = fridgeBasketList.get(i).getIngredientName();
-                String ingredientIcon = fridgeBasketList.get(i).getIngredientIcon();
-                Integer ingredientCategoryIdx = fridgeBasketList.get(i).getIngredientCategoryIdx();
-                String expiredAt = fridgeBasketList.get(i).getExpiredAt();
-                String storageMethod = fridgeBasketList.get(i).getStorageMethod();
-                Integer count = fridgeBasketList.get(i).getCount();
-
-                if (ingredientName == null || ingredientName.length()==0 ) {
+                if (ingredientName == null || ingredientName.length()==0) {
                     return new BaseResponse<>(POST_FRIDGES_EMPTY_INGREDIENT_NAME);
                 }
-                if (ingredientIcon == null || ingredientIcon.length()==0 ) {
+                if (ingredientIcon == null || ingredientIcon.length()==0) {
                     return new BaseResponse<>(POST_FRIDGES_EMPTY_INGREDIENT_ICON);
                 }
-                //int만 입력하도록?
-                if (ingredientCategoryIdx == null || ingredientCategoryIdx<=0 ) {
+                if (ingredientCategoryIdx == null || ingredientCategoryIdx<=0) {
                     return new BaseResponse<>(POST_FRIDGES_DIRECT_BASKET_EMPTY_INGREDIENT_CATEGORY_IDX);
                 }
-                if (storageMethod == null || storageMethod.length()==0 ) {
+                if (storageMethod == null || storageMethod.length()==0) {
                     return new BaseResponse<>(EMPTY_STORAGE_METHOD);
                 }
-                if (count == null || count<=0  ) {
+                if (count == null || count<=0) {
                     return new BaseResponse<>(EMPTY_INGREDIENT_COUNT);
                 }
                 if (expiredAt !=null && expiredAt.length()!=0 && !expiredAt.matches("^\\d{4}\\.(0[1-9]|1[012])\\.(0[1-9]|[12][0-9]|3[01])$")) {
                     return new BaseResponse<>(INVALID_DATE);
                 }
-                Boolean existIngredientName = fridgeProvider.existIngredient(ingredientName,userIdx);
-                if (existIngredientName) {
-                    return new BaseResponse<>(POST_FRIDGES_EXIST_INGREDIENT_NAME,ingredientName);
-                }
-                ArrayList storageMethods = new ArrayList();
-                storageMethods.add("냉장");
-                storageMethods.add("냉동");
-                storageMethods.add("실온");
-
-                if (!storageMethods.contains(storageMethod)){
+                if (!storageMethod.equals("냉장") && !storageMethod.equals("냉동") && !storageMethod.equals("실온")){
                     return new BaseResponse<>(INVALID_STORAGE_METHOD);
                 }
-
             }
 
-            List<PostFridgesRes> postFridgesBasketRes = fridgeService.createFridges(parameters,userIdx);
+            List<PostFridgesRes> postFridgesBasketRes = fridgeService.createFridges(parameters, userIdx);
 
             return new BaseResponse<>(postFridgesBasketRes);
         } catch (BaseException exception) {
-            return new BaseResponse<>(exception.getStatus());
+            return new BaseResponse<>(exception.getStatus(), exception.getIngredientName());
         }
     }
 
@@ -118,15 +104,11 @@ public class FridgeController {
      */
     @GetMapping("/fridges")
     public BaseResponse<GetFridgesRes> getFridges() {
-
         try {
             Integer userIdx = jwtService.getUserId();
-
             GetFridgesRes getFridgesRes = fridgeProvider.retreiveFridges(userIdx);
 
-
             return new BaseResponse<>(getFridgesRes);
-
         } catch (BaseException exception) {
             return new BaseResponse<>(exception.getStatus());
         }
@@ -140,23 +122,14 @@ public class FridgeController {
      */
     @DeleteMapping("/fridges/ingredient")
     public BaseResponse<Void> deleteFridgesIngredient(@RequestBody DeleteFridgesIngredientReq parameters) throws BaseException {
-
-        List<String> ingrdientNames = parameters.getIngredientName();
-
-        if (ingrdientNames == null) {
-            return new BaseResponse<>(EMPTY_INGREDIENT);
-        }
-        Integer userIdx = jwtService.getUserId();
-        for(String ingrdientName : ingrdientNames) {
-            Boolean existIngredient = fridgeProvider.existIngredient(ingrdientName, userIdx);
-            if (!existIngredient) {
-                return new BaseResponse<>(NOT_FOUND_INGREDIENT);
-            }
-        }
-
         try {
+            Integer userIdx = jwtService.getUserId();
+            List<String> ingrdientNames = parameters.getIngredientName();
+            if (ingrdientNames == null || ingrdientNames.isEmpty()) {
+                return new BaseResponse<>(EMPTY_INGREDIENT);
+            }
 
-            fridgeService.deleteFridgeIngredient(userIdx,parameters);
+            fridgeService.deleteFridgeIngredient(userIdx, parameters);
             return new BaseResponse<>(SUCCESS);
         } catch (BaseException exception) {
             return new BaseResponse<>(exception.getStatus());
@@ -173,22 +146,18 @@ public class FridgeController {
     @ResponseBody
     @PatchMapping("/fridges/ingredient")
     public BaseResponse<Void> patchFridgesIngredient(@RequestBody PatchFridgesIngredientReq parameters) throws BaseException {
-
         try {
             Integer userIdx = jwtService.getUserId();
-
             List<PatchFridgeList> patchFridgeList = parameters.getPatchFridgeList();
-
             if (patchFridgeList==null || patchFridgeList.isEmpty()) {
                 return new BaseResponse<>(EMPTY_PATCH_FRIDGE_LIST);
             }
 
-
-            for (int i = 0; i < patchFridgeList.size(); i++) {
-                String ingredientName = patchFridgeList.get(i).getIngredientName();
-                String expiredAt = patchFridgeList.get(i).getExpiredAt();
-                String storageMethod = patchFridgeList.get(i).getStorageMethod();
-                Integer count = patchFridgeList.get(i).getCount();
+            for (PatchFridgeList patchFridge : patchFridgeList) {
+                String ingredientName = patchFridge.getIngredientName();
+                String expiredAt = patchFridge.getExpiredAt();
+                String storageMethod = patchFridge.getStorageMethod();
+                Integer count = patchFridge.getCount();
 
                 if (ingredientName == null || ingredientName.length()==0 ) {
                     return new BaseResponse<>(POST_FRIDGES_EMPTY_INGREDIENT_NAME);
@@ -199,21 +168,12 @@ public class FridgeController {
                 if (count == null || count<=0  ) {
                     return new BaseResponse<>(EMPTY_INGREDIENT_COUNT);
                 }
-                if (expiredAt !=null && ingredientName.length()!=0 && !expiredAt.matches("^\\d{4}\\.(0[1-9]|1[012])\\.(0[1-9]|[12][0-9]|3[01])$")) {
+                if (expiredAt !=null && expiredAt.length()!=0 && !expiredAt.matches("^\\d{4}\\.(0[1-9]|1[012])\\.(0[1-9]|[12][0-9]|3[01])$")) {
                     return new BaseResponse<>(INVALID_DATE);
                 }
-                Boolean existIngredientName = fridgeProvider.existIngredient(ingredientName, userIdx);
-                if (!existIngredientName) {
-                    return new BaseResponse<>(NOT_FOUND_INGREDIENT);
-                }
-                ArrayList storageMethods = new ArrayList();
-                storageMethods.add("냉장");
-                storageMethods.add("냉동");
-                storageMethods.add("실온");
-                if (!storageMethods.contains(storageMethod)){
+                if (!storageMethod.equals("냉장") && !storageMethod.equals("냉동") && !storageMethod.equals("실온")){
                     return new BaseResponse<>(INVALID_STORAGE_METHOD);
                 }
-
             }
 
             fridgeService.updateFridgeIngredient(parameters,userIdx);
@@ -221,6 +181,8 @@ public class FridgeController {
             return new BaseResponse<>(SUCCESS);
         } catch (BaseException exception) {
             return new BaseResponse<>(exception.getStatus());
+        } catch (Exception exception) {
+            throw new BaseException(FAILED_TO_PATCH_FRIDGES_INGREDIENT);
         }
     }
 
@@ -233,10 +195,8 @@ public class FridgeController {
     @ResponseBody
     @GetMapping("/fridges/recipe")
     public BaseResponse<GetFridgesRecipeRes> getFridgesRecipe(@RequestParam(value = "start") Integer start, @RequestParam(value = "display") Integer display)  {
-
         try {
             Integer userIdx = jwtService.getUserId();
-
 
             GetFridgesRecipeRes getFridgesRecipeRes = fridgeProvider.retreiveFridgesRecipe(userIdx, start, display);
